+++ packages/@oak-tree-house/vuepress-plugin-diagrams/event.js
@@ -2,7 +2,9 @@ import Vue from 'vue'
 
 export default new Vue({
   data: () => ({
-    sequenceDiagramsStatus: 0,  // 0 for not started, 1 for loading, 2 for loaded
-    flowchartDiagramsStatus: 0
+    // 0 for not started, 1 for loading, 2 for loaded
+    sequenceDiagramsStatus: 0,
+    flowchartDiagramsStatus: 0,
+    mermaidDiagramsStatus: 0
   })
 })