+++ packages/@oak-tree-house/vuepress-plugin-diagrams/util.js
@@ -11,3 +11,7 @@ export function loadScript (src, id) {
     }
   })
 }
+
+export function generateId () {
+  return '_' + Math.random().toString(36).substr(2, 9)
+}