+++ packages/@oak-tree-house/vuepress-theme-vuetify-blog/components/SidebarToc.vue
@@ -92,7 +92,10 @@ export default {
   },
   methods: {
     updateTocListener () {
-      this.$nextTick(() => this.updateToc())
+      this.$nextTick(() => {
+        this.updateToc()
+        this.updateTocOpenAndActive()
+      })
     },
     resolveHeaders,
     updateToc () {